Influunt.MongoStorage splited to 2 modules: 1) Influunt.Storage - service implementations using IRepository 2) Influunt.Storage.Mongo - IRepository implemetation via MongoDriver
Updated Readme.md:
1) Added screenshots
