+++ src/Influunt.Storage.Mongo/Abstractions/IMongoStorageConfiguration.cs
@@ -1,4 +1,4 @@
-﻿namespace Influunt.MongoStorage.Abstractions
+﻿namespace Influunt.Storage.Mongo.Abstractions
 {
     public interface IMongoStorageConfiguration
     {
