+++ src/Influunt.Storage.Mongo/Influunt.Storage.Mongo.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.7" />
+    <PackageReference Include="MongoDB.Driver" Version="2.11.4" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Influunt.Feed\Influunt.Feed.csproj" />
+  </ItemGroup>
+</Project>
