+++ src/Influunt.Storage/Entity/FavoriteFeedItemMap.cs
@@ -1,6 +1,7 @@
 ﻿using Influunt.Feed.Entity;
+using Influunt.Storage.Mongo;
 
-namespace Influunt.MongoStorage.Entity
+namespace Influunt.Storage.Entity
 {
     public class FavoriteFeedItemMap: EntityMapClass<FavoriteFeedItem>
     {
