+++ src/Influunt.Storage/Entity/FeedChannelMap.cs
@@ -1,6 +1,7 @@
 ﻿using Influunt.Feed.Entity;
+using Influunt.Storage.Mongo;
 
-namespace Influunt.MongoStorage.Entity
+namespace Influunt.Storage.Entity
 {
     public class FeedChannelMap: EntityMapClass<FeedChannel>
     {
