+++ src/Influunt.Storage/Influunt.Storage.csproj
@@ -12,10 +12,10 @@
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.7" />
     <PackageReference Include="Microsoft.AspNetCore.DataProtection" Version="3.1.7" />
-    <PackageReference Include="MongoDB.Driver" Version="2.11.4" />
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Influunt.Storage.Mongo\Influunt.Storage.Mongo.csproj" />
     <ProjectReference Include="..\Influunt.Feed\Influunt.Feed.csproj" />
   </ItemGroup>
