+++ src/Influunt.Storage/Services/FavoriteService.cs
@@ -4,10 +4,10 @@ using System.Linq;
 using System.Threading.Tasks;
 using Influunt.Feed;
 using Influunt.Feed.Entity;
-using Influunt.MongoStorage.Abstractions;
+using Influunt.Storage.Mongo.Abstractions;
 using MongoDB.Bson;
 
-namespace Influunt.MongoStorage.Services
+namespace Influunt.Storage.Services
 {
     public class FavoriteService: IFavoriteFeedService
     {
